nifti

Enum Endianness

pub enum Endianness {
    Little,
    Big,
}

Enumerate for materializing the two kinds of machine byte order supported by Rust in a dynamic fashion. That is, the information of whether to read or write data in Little Endian or in Big Endian is resolved at run time by observing this value.

Using this type as the generic endianness type E in a ByteOrdered is useful when this information can only be retrieved from a source that is unknown to the compiler.

Variants

Little

Little Endian

Big

Big Endian

Implementations

impl Endianness

pub fn native() -> Endianness

Obtains this system’s native endianness.

On this platform, the function returns Endianness::Little.

pub fn le_iff(e: bool) -> Endianness

Obtains Little Endian if and only if the given value is true.

Examples

let data: &[u8] = &[4, 1];
let e = Endianness::le_iff(2 + 2 == 4);
assert_eq!(e.read_u16(data).unwrap(), 260);

let e = Endianness::le_iff(2 + 2 >= 5);
assert_eq!(e.read_u16(data).unwrap(), 1025);

pub fn be_iff(e: bool) -> Endianness

Obtains Big Endian if and only if the given value is true.

Examples

assert_eq!(Endianness::be_iff(2 + 2 == 4), Endianness::Big);
assert_eq!(Endianness::be_iff(2 + 2 >= 5), Endianness::Little);

pub fn to_opposite(self) -> Endianness

Obtains the opposite endianness: Little Endian returns Big Endian and vice versa.