nifti

Module extension

This module contains definitions for the extension and related types. Extensions are optional data frames sitting before the voxel data. When present, an extender frame of 4 bytes is also present at the end of the NIFTI-1 header, with the first byte set to something other than 0.

Structs

Extender

Data type for the extender code.

Extension

Data type for the raw contents of an extension. Users of this type have to reinterpret the data to suit their needs.

ExtensionSequence

Data type for aggregating the extender code and all extensions.