nifti

Module object

Module for handling and retrieving complete NIFTI-1 objects.

Structs

GenericNiftiObject

Generic data type for a NIfTI object.

ReaderOptions

Options and flags which can be used to configure how a NIfTI image is read.

ReaderStreamedOptions

Options and flags which can be used to configure how a NIfTI image is read and iterated.

Traits

NiftiObject

Trait type for all possible implementations of owning NIFTI-1 objects. Objects contain a NIFTI header, a volume, and a possibly empty extension sequence.

Type Aliases

GzDecodedFile

A reader for a GZip encoded file.

InMemNiftiObject

A NIfTI object containing an in-memory volume.

MaybeGzDecodedFile

A reader for a file which might be GZip encoded based on some run-time condition.

StreamedNiftiObject

A NIfTI object containing a streamed volume.