nifti::object

Trait NiftiObject

pub trait NiftiObject {
    type Volume: NiftiVolume;

    // Required methods
    fn header(&self) -> &NiftiHeader;
    fn header_mut(&mut self) -> &mut NiftiHeader;
    fn extensions(&self) -> &ExtensionSequence;
    fn volume(&self) -> &Self::Volume;
    fn into_volume(self) -> Self::Volume;
}

Trait type for all possible implementations of owning NIFTI-1 objects. Objects contain a NIFTI header, a volume, and a possibly empty extension sequence.

Required Associated Types

type Volume: NiftiVolume

The concrete type of the volume.

Required Methods

fn header(&self) -> &NiftiHeader

Obtain a reference to the NIFTI header.

fn header_mut(&mut self) -> &mut NiftiHeader

Obtain a mutable reference to the NIFTI header.

fn extensions(&self) -> &ExtensionSequence

Obtain a reference to the object’s extensions.

fn volume(&self) -> &Self::Volume

Obtain a reference to the object’s volume.

fn into_volume(self) -> Self::Volume

Move the volume out of the object, discarding the header and extensions.