nifti

Module typedef

This module contains multiple types defined by the standard. At the moment, not all of them are used internally (NiftiType makes the exception, which also provides a safe means of reading voxel values). However, primitive integer values can be converted to these types and vice-versa.

Enums

Intent

An enum type for representing a NIFTI intent code.

NiftiType

Data type for representing a NIFTI value type in a volume. Methods for reading values of that type from a source are also included.

SliceOrder

An enum type for representing the slice order.

Unit

An enum type which represents a unit type.

XForm

An enum type for representing a NIFTI XForm.