nifti::typedef

Enum Intent

#[repr(u16)]
pub enum Intent {

    None = 0,
    Correl = 2,
    Ttest = 3,
    Ftest = 4,
    Zscore = 5,
    Chisq = 6,
    Beta = 7,
    Binom = 8,
    Gamma = 9,
    Poisson = 10,
    Normal = 11,
    FtestNonc = 12,
    ChisqNonc = 13,
    Logistic = 14,
    Laplace = 15,
    Uniform = 16,
    TtestNonc = 17,
    Weibull = 18,
    Chi = 19,
    Invgauss = 20,
    Extval = 21,
    Pval = 22,
    Logpval = 23,
    Log10pval = 24,
    Estimate = 1_001,
    Label = 1_002,
    Neuroname = 1_003,
    Genmatrix = 1_004,
    Symmatrix = 1_005,
    Dispvect = 1_006,
    Vector = 1_007,
    Pointset = 1_008,
    Triangle = 1_009,
    Quaternion = 1_010,
    Dimless = 1_011,
    TimeSeries = 2_001,
    NodeIndex = 2_002,
    RgbVector = 2_003,
    RgbaVector = 2_004,
    Shape = 2_005,
    FslFnirtDisplacementField = 2_006,
    FslCubicSplineCoefficients = 2_007,
    FslDctCoefficients = 2_008,
    FslQuadraticSplineCoefficients = 2_009,
    FslTopupCubicSplineCoefficients = 2_016,
    FslTopupQuadraticSplineCoefficients = 2_017,
    FslTopupField = 2_018,

}

An enum type for representing a NIFTI intent code.

Variants

None = 0

default: no intention is indicated in the header.

Correl = 2

nifti1 intent codes, to describe intended meaning of dataset contents

Ttest = 3

[C2, chap 28] Student t statistic (1 param): p1 = DOF.

Ftest = 4

[C2, chap 27] Fisher F statistic (2 params): p1 = numerator DOF, p2 = denominator DOF.

Zscore = 5

[C1, chap 13] Standard normal (0 params): Density = N(0,1).

Chisq = 6

[C1, chap 18] Chi-squared (1 param): p1 = DOF. Density(x) proportional to exp(-x/2) * x^(p1/2-1).

Beta = 7

[C2, chap 25] Beta distribution (2 params): p1=a, p2=b. Density(x) proportional to x^(a-1) * (1-x)^(b-1).

Binom = 8

[U, chap 3] Binomial distribution (2 params): p1 = number of trials, p2 = probability per trial. Prob(x) = (p1 choose x) * p2^x * (1-p2)^(p1-x), for x=0,1,...,p1.

Gamma = 9

[C1, chap 17] Gamma distribution (2 params): p1 = shape, p2 = scale. Density(x) proportional to x^(p1-1) * exp(-p2*x).

Poisson = 10

[U, chap 4] Poisson distribution (1 param): p1 = mean. Prob(x) = exp(-p1) * p1^x / x! , for x=0,1,2,....

Normal = 11

[C1, chap 13] Normal distribution (2 params): p1 = mean, p2 = standard deviation.

FtestNonc = 12

[C2, chap 30] Noncentral F statistic (3 params): p1 = numerator DOF, p2 = denominator DOF, p3 = numerator noncentrality parameter.

ChisqNonc = 13

[C2, chap 29] Noncentral chi-squared statistic (2 params): p1 = DOF, p2 = noncentrality parameter.

Logistic = 14

[C2, chap 23] Logistic distribution (2 params): p1 = location, p2 = scale. Density(x) proportional to sech^2((x-p1)/(2*p2)).

Laplace = 15

[C2, chap 24] Laplace distribution (2 params): p1 = location, p2 = scale. Density(x) proportional to exp(-abs(x-p1)/p2).

Uniform = 16

[C2, chap 26] Uniform distribution: p1 = lower end, p2 = upper end.

TtestNonc = 17

[C2, chap 31] Noncentral t statistic (2 params): p1 = DOF, p2 = noncentrality parameter.

Weibull = 18

[C1, chap 21] Weibull distribution (3 params): p1 = location, p2 = scale, p3 = power. Density(x) proportional to ((x-p1)/p2)^(p3-1) * exp(-((x-p1)/p2)^p3) for x > p1.

Chi = 19

[C1, chap 18] Chi distribution (1 param): p1 = DOF. Density(x) proportional to x^(p1-1) * exp(-x^2/2) for x > 0. p1 = 1 = ‘half normal’ distribution p1 = 2 = Rayleigh distribution p1 = 3 = Maxwell-Boltzmann distribution.

Invgauss = 20

[C1, chap 15] Inverse Gaussian (2 params): p1 = mu, p2 = lambda Density(x) proportional to exp(-p2*(x-p1)^2/(2*p1^2*x)) / x^3 for x > 0.

Extval = 21

[C2, chap 22] Extreme value type I (2 params): p1 = location, p2 = scale cdf(x) = exp(-exp(-(x-p1)/p2)).

Pval = 22

Data is a ‘p-value’ (no params).

Logpval = 23

Data is ln(p-value) (no params). To be safe, a program should compute p = exp(-abs(this_value)). The nifti_stats.c library returns this_value as positive, so that this_value = -log(p).

Log10pval = 24

Data is log10(p-value) (no params). To be safe, a program should compute p = pow(10.,-abs(this_value)). The nifti_stats.c library returns this_value as positive, so that this_value = -log10(p).

Estimate = 1_001

To signify that the value at each voxel is an estimate of some parameter, set intent_code = NIFTI_INTENT_ESTIMATE. The name of the parameter may be stored in intent_name.

Label = 1_002

To signify that the value at each voxel is an index into some set of labels, set intent_code = NIFTI_INTENT_LABEL. The filename with the labels may stored in aux_file.

Neuroname = 1_003

To signify that the value at each voxel is an index into the NeuroNames labels set, set intent_code = NIFTI_INTENT_NEURONAME.

Genmatrix = 1_004

To store an M x N matrix at each voxel:

dataset must have a 5th dimension (dim[0]=5 and dim[5]>1)

intent_code must be NIFTI_INTENT_GENMATRIX

dim[5] must be M*N

intent_p1 must be M (in float format)

intent_p2 must be N (ditto)

the matrix values A[i][j] are stored in row-order:

A[0][0] A[0][1] ... A[0][N-1]

A[1][0] A[1][1] ... A[1][N-1]

etc., until

A[M-1][0] A[M-1][1] ... A[M-1][N-1]

Symmatrix = 1_005

To store an NxN symmetric matrix at each voxel:

dataset must have a 5th dimension

intent_code must be NIFTI_INTENT_SYMMATRIX

dim[5] must be N*(N+1)/2

intent_p1 must be N (in float format)

the matrix values A[i][j] are stored in row-order:

A[0][0]

A[1][0] A[1][1]

A[2][0] A[2][1] A[2][2]

etc.: row-by-row

Dispvect = 1_006

To signify that the vector value at each voxel is to be taken as a displacement field or vector:

dataset must have a 5th dimension

intent_code must be NIFTI_INTENT_DISPVECT

dim[5] must be the dimensionality of the displacment vector (e.g., 3 for spatial displacement, 2 for in-plane)

(specifically for displacements)

Vector = 1_007

(for any other type of vector)

Pointset = 1_008

To signify that the vector value at each voxel is really a spatial coordinate (e.g., the vertices or nodes of a surface mesh):

dataset must have a 5th dimension

intent_code must be NIFTI_INTENT_POINTSET

dim[0] = 5

dim[1] = number of points

dim[2] = dim[3] = dim[4] = 1

dim[5] must be the dimensionality of space (e.g., 3 => 3D space).

intent_name may describe the object these points come from (e.g., “pial”, “gray/white” , “EEG”, “MEG”).

Triangle = 1_009

To signify that the vector value at each voxel is really a triple of indexes (e.g., forming a triangle) from a pointset dataset:

dataset must have a 5th dimension

intent_code must be NIFTI_INTENT_TRIANGLE

dim[0] = 5

dim[1] = number of triangles

dim[2] = dim[3] = dim[4] = 1

dim[5] = 3

datatype should be an integer type (preferably NiftiType::Int32)

the data values are indexes (0,1,…) into a pointset dataset.

Quaternion = 1_010

To signify that the vector value at each voxel is a quaternion:

dataset must have a 5th dimension

intent_code must be NIFTI_INTENT_QUATERNION

dim[0] = 5

dim[5] = 4

datatype should be a floating point type

Dimless = 1_011

Dimensionless value - no params - although, as in _ESTIMATE the name of the parameter may be stored in intent_name.

TimeSeries = 2_001

To signify that the value at each location is from a time series.

NodeIndex = 2_002

To signify that the value at each location is a node index, from a complete surface dataset.

RgbVector = 2_003

To signify that the vector value at each location is an RGB triplet, of whatever type.

dataset must have a 5th dimension

dim[0] = 5

dim[1] = number of nodes

dim[2] = dim[3] = dim[4] = 1

dim[5] = 3

RgbaVector = 2_004

To signify that the vector value at each location is a 4 valued RGBA vector, of whatever type.

dataset must have a 5th dimension

dim[0] = 5

dim[1] = number of nodes

dim[2] = dim[3] = dim[4] = 1

dim[5] = 4

Shape = 2_005

To signify that the value at each location is a shape value, such as the curvature.

FslFnirtDisplacementField = 2_006

FSL FNIRT Displacement field.

FslCubicSplineCoefficients = 2_007

FSL Cubic spline coefficients.

FslDctCoefficients = 2_008

FSL Discrete cosine transform coefficients.

FslQuadraticSplineCoefficients = 2_009

FSL Quadratic spline coefficients.

FslTopupCubicSplineCoefficients = 2_016

FSL Topup cubic spline coefficients.

FslTopupQuadraticSplineCoefficients = 2_017

FSL Topup quadratic spline coefficients.

FslTopupField = 2_018

FSL Topup field.

Implementations

impl Intent

pub fn is_statcode(self) -> bool

Check whether this intent code is used for statistics.