nifti::volume::element

Trait DataElement

pub trait DataElement:
    'static
    + Sized
    + Copy {
    type DataRescaler: NiftiDataRescaler<Self>;

    const DATA_TYPE: NiftiType;

    // Required methods
    fn from_raw<R, E>(src: R, endianness: E) -> Result<Self>
       where R: Read,
             E: Endian;
    fn from_raw_vec_validated<E>(
        vec: Vec<u8>,
        endianness: E,
        datatype: NiftiType,
    ) -> Result<Vec<Self>>
       where E: Endian;

    // Provided methods
    fn from_u8(_value: u8) -> Self { ... }
    fn from_i8(_value: i8) -> Self { ... }
    fn from_u16(_value: u16) -> Self { ... }
    fn from_i16(_value: i16) -> Self { ... }
    fn from_u32(_value: u32) -> Self { ... }
    fn from_i32(_value: i32) -> Self { ... }
    fn from_u64(_value: u64) -> Self { ... }
    fn from_i64(_value: i64) -> Self { ... }
    fn from_f32(_value: f32) -> Self { ... }
    fn from_f64(_value: f64) -> Self { ... }
    fn from_complex32(_value: Complex32) -> Self { ... }
    fn from_complex64(_value: Complex64) -> Self { ... }
    fn from_raw_vec<E>(vec: Vec<u8>, endianness: E) -> Result<Vec<Self>>
       where E: Endian + Clone { ... }

}

Trait type for characterizing a NIfTI data element, implemented for primitive numeric types which are used by the crate to represent voxel values.

Required Associated Constants

const DATA_TYPE: NiftiType

The datatype mapped to the type T

Required Associated Types

type DataRescaler: NiftiDataRescaler<Self>

Implement rescaling for the given data type

Required Methods

fn from_raw<R, E>(src: R, endianness: E) -> Result<Self>
where R: Read, E: Endian,

Read a single element from the given byte source.

fn from_raw_vec_validated<E>( vec: Vec<u8>, endianness: E, datatype: NiftiType, ) -> Result<Vec<Self>>
where E: Endian,

Return a vector of data elements of the native type indicated in the Nifti file with runtime check

Provided Methods

fn from_u8(_value: u8) -> Self

Create a single element by converting a scalar value.

fn from_i8(_value: i8) -> Self

Create a single element by converting a scalar value.

fn from_u16(_value: u16) -> Self

Create a single element by converting a scalar value.

fn from_i16(_value: i16) -> Self

Create a single element by converting a scalar value.

fn from_u32(_value: u32) -> Self

Create a single element by converting a scalar value.

fn from_i32(_value: i32) -> Self

Create a single element by converting a scalar value.

fn from_u64(_value: u64) -> Self

Create a single element by converting a scalar value.

fn from_i64(_value: i64) -> Self

Create a single element by converting a scalar value.

fn from_f32(_value: f32) -> Self

Create a single element by converting a scalar value.

fn from_f64(_value: f64) -> Self

Create a single element by converting a scalar value.

fn from_complex32(_value: Complex32) -> Self

Create a single element by converting a complex value

fn from_complex64(_value: Complex64) -> Self

Create a single element by converting a complex value

fn from_raw_vec<E>(vec: Vec<u8>, endianness: E) -> Result<Vec<Self>>
where E: Endian + Clone,

Transform the given data vector into a vector of data elements.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.