nifti::volume::element

Trait NiftiDataRescaler

pub trait NiftiDataRescaler<T: 'static + Copy> {
    // Required method
    fn nifti_rescale(value: T, slope: f32, intercept: f32) -> T;

    // Provided methods
    fn nifti_rescale_many(value: &[T], slope: f32, intercept: f32) -> Vec<T> { ... }
    fn nifti_rescale_many_inline(value: &mut [T], slope: f32, intercept: f32) { ... }
}

NiftiDataRescaler, a trait for rescaling data elements according to the Nifti 1.1 specification

Required Methods

fn nifti_rescale(value: T, slope: f32, intercept: f32) -> T

Rescale a single value with the given slope and intercept.

Provided Methods

fn nifti_rescale_many(value: &[T], slope: f32, intercept: f32) -> Vec<T>

Rescale a slice of values, with the given slope and intercept.

fn nifti_rescale_many_inline(value: &mut [T], slope: f32, intercept: f32)

Rescale a slice of values inline (inplace), with the given slope and intercept.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.