nifti::volume

Trait NiftiVolume

pub trait NiftiVolume {
    // Required methods
    fn dim(&self) -> &[u16];
    fn data_type(&self) -> NiftiType;

    // Provided method
    fn dimensionality(&self) -> usize { ... }
}

Public API for a NIFTI volume.

This API is currently experimental and will likely be subjected to various changes and additions in future versions.

Required Methods

fn dim(&self) -> &[u16]

Get the dimensions of the volume. Unlike how NIFTI-1 stores dimensions, the returned slice does not include dim[0] and is clipped to the effective number of dimensions.

fn data_type(&self) -> NiftiType

Get this volume’s data type.

Provided Methods

fn dimensionality(&self) -> usize

Get the volume’s number of dimensions. In a fully compliant file, this is equivalent to the corresponding header’s dim[0] field (with byte swapping already applied).